fix(train): five gate failures from the #9976/#9977 stacks
- copying.rs reached 2066 lines. The remembered-set scan and pinned-young
  preflight move to copying/remembered_scan.rs. Their `pub(super)` meant
  `gc` in the parent and means `copying` in the child, so both are widened
  to `pub(in crate::gc)` to keep the original reach.

- regex_census.rs: `rows` is extended only under `regex-engine`, so the
  binding is unused-mut without that feature and REQUIRES mut with it.
  Scoped the allow to the feature-off build rather than dropping `mut`,
  which breaks the feature-on build. (My first attempt dropped it.)

- shapes.rs: #9976 deliberately removed `family_push_back`'s production
  rekey caller — the scanner-internal rekey note explains why re-entering
  the writer funnel mid-walk is wrong — leaving shapes_test_support as its
  only consumer. Gated `#[cfg(test)]` to match.

- Five new holders classified: BOX_YOUNG_ROOTS is covered_elsewhere, since
  every address in that minor remembered set is also in the box registry
  that scan_box_roots_mut walks; the four test seams and the histogram
  counter are not_a_gc_pointer.

- PASS1_MARKED re-audited: gc/mod.rs gains two module declarations and one
  reg_scanner! registration, gc/census.rs widens side_tables() and adds
  census rows plus a test. A scanner registration adds a root SOURCE and
  runs nowhere between the census boundaries; census reporting runs from
  the diagnostic dump, not inside a cycle. LAYOUT_DIAG's entry is deleted:
  the holder became covered, which the gate calls the receipt.
